feat: add stat components for funding rate
