+++ apps/trading/components/market-header/market-header-perp.tsx
@@ -1,23 +1,5 @@
-import { DocsLinks } from '@vegaprotocol/environment';
-import { ExternalLink } from '@vegaprotocol/ui-toolkit';
 import type { Market } from '@vegaprotocol/markets';
-import { addDecimalsFormatNumber, fromNanoSeconds } from '@vegaprotocol/utils';
-import {
-  getAsset,
-  getDataSourceSpecForSettlementSchedule,
-  isMarketInAuction,
-  marketInfoProvider,
-  useFundingPeriodsQuery,
-  useFundingRate,
-  useMarketTradingMode,
-  useExternalTwap,
-  getQuoteName,
-  useMarketState,
-} from '@vegaprotocol/markets';
-import { HeaderStat } from '../../components/header';
-import { useEffect, useState } from 'react';
-import { useDataProvider } from '@vegaprotocol/data-provider';
-import { PriceCell } from '@vegaprotocol/datagrid';
+import { getAsset, getQuoteName } from '@vegaprotocol/markets';
 import { useT } from '../../lib/use-t';
 import * as Stats from './stats';
 
@@ -31,7 +13,6 @@ export const MarketHeaderPerp = ({ market }: MarketHeaderPerpProps) => {
   }
 
   const t = useT();
-
   const asset = getAsset(market);
   const quoteUnit = getQuoteName(market);
 
@@ -57,10 +38,7 @@ export const MarketHeaderPerp = ({ market }: MarketHeaderPerpProps) => {
         positionDecimalPlaces={market.positionDecimalPlaces}
         quoteUnit={quoteUnit}
       />
-      <Stats.MarketTradingModeStat
-        marketId={market.id}
-        initialTradingMode={market.tradingMode}
-      />
+      <Stats.MarketTradingModeStat marketId={market.id} />
       <Stats.MarketStateStat market={market} />
       <Stats.AssetStat
         heading={t('Settlement asset')}
@@ -72,146 +50,11 @@ export const MarketHeaderPerp = ({ market }: MarketHeaderPerpProps) => {
         assetDecimals={asset.decimals}
         quantum={asset.quantum}
       />
-      <HeaderStat
-        heading={`${t('Funding Rate')} / ${t('Countdown')}`}
-        data-testid="market-funding"
-      >
-        <div className="flex gap-2">
-          <FundingRate marketId={market.id} />
-          <FundingCountdown marketId={market.id} />
-        </div>
-      </HeaderStat>
-      <HeaderStat
-        heading={`${t('Index Price')}`}
-        description={
-          <div className="p1">
-            {t(
-              'The external time weighted average price (TWAP) received from the data source defined in the data sourcing specification.'
-            )}
-            {DocsLinks && (
-              <ExternalLink href={DocsLinks.ETH_DATA_SOURCES} className="mt-2">
-                {t('Find out more')}
-              </ExternalLink>
-            )}
-          </div>
-        }
-        data-testid="index-price"
-      >
-        <IndexPrice
-          marketId={market.id}
-          decimalPlaces={
-            market.tradableInstrument.instrument.product.settlementAsset
-              .decimals
-          }
-        />
-      </HeaderStat>
+      <Stats.FundingRate marketId={market.id} />
+      <Stats.IndexPriceStat
+        marketId={market.id}
+        assetDecimals={asset.decimals}
+      />
     </>
-  );
-};
-
-export const FundingRate = ({ marketId }: { marketId: string }) => {
-  const { data: fundingRate } = useFundingRate(marketId);
-  return (
-    <div data-testid="funding-rate">
-      {fundingRate ? `${(Number(fundingRate) * 100).toFixed(4)}%` : '-'}
-    </div>
-  );
-};
-
-export const IndexPrice = ({
-  marketId,
-  decimalPlaces,
-}: {
-  marketId: string;
-  decimalPlaces?: number;
-}) => {
-  const { data: externalTwap } = useExternalTwap(marketId);
-  return externalTwap && decimalPlaces ? (
-    <PriceCell
-      value={Number(externalTwap)}
-      valueFormatted={addDecimalsFormatNumber(externalTwap, decimalPlaces)}
-    />
-  ) : (
-    '-'
-  );
-};
-
-const useNow = () => {
-  const [now, setNow] = useState(Date.now());
-  useEffect(() => {
-    const interval = setInterval(() => setNow(Date.now()), 1000);
-    return () => clearInterval(interval);
-  }, []);
-  return now;
-};
-
-const useEvery = (marketId: string, skip: boolean) => {
-  const { data: marketInfo } = useDataProvider({
-    dataProvider: marketInfoProvider,
-    variables: { marketId },
-    skip,
-  });
-  let every: number | undefined = undefined;
-  const sourceType =
-    marketInfo &&
-    getDataSourceSpecForSettlementSchedule(
-      marketInfo.tradableInstrument.instrument.product
-    )?.data.sourceType.sourceType;
-
-  if (sourceType?.__typename === 'DataSourceSpecConfigurationTimeTrigger') {
-    every = sourceType.triggers?.[0]?.every ?? undefined;
-    if (every) {
-      every *= 1000;
-    }
-  }
-  return every;
-};
-
-const useStartTime = (marketId: string, skip: boolean) => {
-  const { data: fundingPeriods } = useFundingPeriodsQuery({
-    pollInterval: 5000,
-    skip,
-    variables: {
-      marketId: marketId,
-      pagination: { first: 1 },
-    },
-  });
-  const node = fundingPeriods?.fundingPeriods.edges?.[0]?.node;
-  let startTime: number | undefined = undefined;
-  if (node && node.startTime && !node.endTime) {
-    startTime = fromNanoSeconds(node.startTime).getTime();
-  }
-  return startTime;
-};
-
-const padStart = (n: number) => n.toString().padStart(2, '0');
-
-const useFormatCountdown = (
-  now: number,
-  startTime?: number,
-  every?: number
-) => {
-  const t = useT();
-  if (startTime && every) {
-    const diff = every - ((now - startTime) % every);
-    const hours = (diff / 3.6e6) | 0;
-    const mins = ((diff % 3.6e6) / 6e4) | 0;
-    const secs = Math.round((diff % 6e4) / 1e3);
-    return `${padStart(hours)}:${padStart(mins)}:${padStart(secs)}`;
-  }
-  return t('Unknown');
-};
-
-export const FundingCountdown = ({ marketId }: { marketId: string }) => {
-  const now = useNow();
-  const { data: marketTradingMode } = useMarketTradingMode(marketId);
-  const skip = !marketTradingMode || isMarketInAuction(marketTradingMode);
-  const startTime = useStartTime(marketId, skip);
-  const every = useEvery(marketId, skip);
-
-  return (
-    <div data-testid="funding-countdown">
-      {useFormatCountdown(now, startTime, every)}
-    </div>
   );
 };
