+++ apps/trading/components/market-header/stats/index.ts
@@ -7,3 +7,5 @@ export { ExpiryStat } from './expiry';
 export { MarketStateStat } from './market-state';
 export { MarketTradingModeStat } from './market-trading-mode';
 export { LiquidityStat } from './liquidity';
+export { IndexPriceStat } from './index-price';
+export { FundingRate } from './funding-rate';
